requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyngrok==7.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
Werkzeug==3.0.1
zipp==3.17.0
